LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/unhandled-promise-rejections/promise-rejection-events-attached-in-event-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: e

PASS Attaching a handler in unhandledrejection should not trigger rejectionhandled.
